jbot/services/discord_bot/requirements.txt
async-timeout>=4.0.2
python-dotenv==1.0.0
discord.py>=2.0.0
aiohttp>=3.8.0
pynacl>=1.4.0
yt-dlp>=2023.3.4